# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RECALL API
  version: 0.1.0
  description: 个性化游戏生成与分享平台 MVP API。
servers:
  - url: http://127.0.0.1:8080
    description: Local development
paths:
...
  /api/v1/games/{gameId}/share-links:
    parameters:
      - $ref: "#/components/parameters/GameID"
    get:
      operationId: listShareLinks
      summary: 获取游戏的分享历史及可再次复制的链接
      tags: [Share Links]
      responses:
        "200":
          description: 分享链接列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLinkListResponse"
        "404":
          $ref: "#/components/responses/APIError"
    post:
      operationId: createShareLink
      summary: 为创建完成的当前版本创建分享链接
      tags: [Share Links]
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/Origin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShareLinkRequest"
      responses:
        "201":
          description: 新分享链接；Secret 只出现在 URL Fragment 中
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLinkResponse"
        "403":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
  /api/v1/games/{gameId}/share-links/{shareId}:
    parameters:
      - $ref: "#/components/parameters/GameID"
      - $ref: "#/components/parameters/ShareID"
    get:
      operationId: getShareLink
      summary: 获取单个制作方分享链接
      tags: [Share Links]
      responses:
        "200":
          description: 分享链接
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLinkResponse"
        "404":
          $ref: "#/components/responses/APIError"
    delete:
      operationId: revokeShareLink
      summary: 主动停止分享并立即使相关游玩会话失效
      tags: [Share Links]
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/Origin"
      responses:
        "200":
          description: 已停止的分享链接
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLinkResponse"
        "404":
          $ref: "#/components/responses/APIError"
  /api/v1/public/shares/{publicId}/resolve:
    parameters:
      - $ref: "#/components/parameters/PublicID"
    post:
      operationId: resolvePublicShare
      summary: 使用 Fragment 中的 Secret 解析公开分享开场信息
      tags: [Public Play]
      parameters:
        - $ref: "#/components/parameters/Origin"
      requestBody:
        $ref: "#/components/requestBodies/ShareSecret"
      responses:
        "200":
          description: 不包含账号、原始素材和创建输入的分享信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicShareResponse"
        "410":
          $ref: "#/components/responses/APIError"
  /api/v1/public/shares/{publicId}/play-sessions:
    parameters:
      - $ref: "#/components/parameters/PublicID"
    post:
      operationId: createPublicPlaySession
      summary: 在有效分享上开始最长 30 分钟的一局游戏
      tags: [Public Play]
      parameters:
        - $ref: "#/components/parameters/Origin"
      requestBody:
        $ref: "#/components/requestBodies/ShareSecret"
      responses:
        "201":
          description: 游玩会话；不透明凭据写入独立 HttpOnly Cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaySessionResponse"
        "410":
          $ref: "#/components/responses/APIError"
  /api/v1/public/play-sessions/current/game-config:
    get:
      operationId: getPublicGameConfig
      summary: 使用游玩会话读取经过 Schema 校验的固定模板配置
      tags: [Public Play]
      responses:
        "200":
          description: 固定模板配置和允许公开的短期资源地址
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameConfigResponse"
        "401":
          $ref: "#/components/responses/APIError"
  /api/v1/public/play-sessions/current/refresh-assets:
    post:
      operationId: refreshPublicAssets
      summary: 为当前游玩会话刷新允许公开的短期资源地址
      tags: [Public Play]
      parameters:
        - $ref: "#/components/parameters/Origin"
      responses:
        "200":
          description: 更新后的公开资源地址
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicAssetListResponse"
        "401":
          $ref: "#/components/responses/APIError"
  /api/v1/public/play-sessions/current/events:
    post:
      operationId: recordPublicPlayBehaviorEvent
      summary: 上报当前有效公开游玩会话的白名单前端事件
      description: 仅 play/all Surface 挂载；第一版只接受 play.completed 和 play.replayed，所有业务关联均由 HttpOnly 游玩会话推导。
      tags: [Analytics, Public Play]
      parameters:
        - $ref: "#/components/parameters/Origin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicPlayAnalyticsEventRequest"
      responses:
        "200":
          description: 合法幂等重试，返回原事件
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsEventWriteResponse"
        "201":
          description: 新事件已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsEventWriteResponse"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "429":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
components:
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
    Origin:
      name: Origin
      in: header
      required: true
      schema:
        type: string
        format: uri
      example: http://127.0.0.1:5173
    GameID:
      name: gameId
      in: path
      required: true
      schema:
        type: string
...
    ShareID:
      name: shareId
      in: path
      required: true
      schema:
        type: string
    PublicID:
      name: publicId
      in: path
      required: true
      description: 公开查询标识，不单独构成授权凭据。
      schema:
        type: string
  requestBodies:
    ShareSecret:
      required: true
      description: Secret 必须来自 URL Fragment，不得放入 Query String。
      content:
        application/json:
          schema:
            type: object
            required: [secret]
            properties:
              secret:
                type: string
  responses:
    APIError:
      description: 请求失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    CreateShareLinkRequest:
      type: object
      required: [expiresAt]
      properties:
        expiresAt:
          type: string
          format: date-time
          description: 必须晚于当前时间，且不能超过创建时间后 90 天。
    ShareLink:
      type: object
      required: [id, gameId, gameVersionId, publicId, url, status, expiresAt, revokedAt, createdAt]
      properties:
        id:
          type: string
        gameId:
          type: string
        gameVersionId:
          type: string
        publicId:
          type: string
        url:
          type: string
          description: Secret 只存在于 URL Fragment；仅返回给游戏所有者。
        status:
          type: string
          enum: [active, expired, revoked]
        expiresAt:
          type: string
          format: date-time
        revokedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ShareLinkResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: "#/components/schemas/ShareLink"
        requestId:
          type: string
    ShareLinkListResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ShareLink"
        requestId:
          type: string
    PublicShareResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          type: object
          required: [creator, share, game]
          properties:
            creator:
              type: object
              required: [displayName]
              properties:
                displayName:
                  type: string
            share:
              type: object
              required: [expiresAt]
              properties:
                expiresAt:
                  type: string
                  format: date-time
            game:
              type: object
              required: [title, ready]
              properties:
                title:
                  type: string
                ready:
                  type: boolean
        requestId:
          type: string
    PlaySessionResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          type: object
          required: [expiresAt, game]
          properties:
            expiresAt:
              type: string
              format: date-time
            game:
              type: object
              required: [title, templateId, templateVersion]
              properties:
                title:
                  type: string
                templateId:
                  type: string
                  enum: [memory-game, love-journey]
                templateVersion:
                  type: string
                  enum: [1.0.0, 1.1.0]
        requestId:
          type: string
    PublicAsset:
      type: object
      required: [key, type, url, mimeType, expiresAt]
      properties:
        key:
          type: string
        type:
          type: string
          enum: [image]
        url:
          type: string
          format: uri
        mimeType:
          type: string
        expiresAt:
          type: string
          format: date-time
    GameConfigResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          type: object
          required: [templateId, templateVersion, configVersion, config, assets, playSessionExpiresAt]
          properties:
            templateId:
              type: string
              enum: [memory-game, love-journey]
            templateVersion:
              type: string
              enum: [1.0.0, 1.1.0]
            configVersion:
              type: integer
              enum: [1]
            config:
              type: object
              additionalProperties: false
              required: [openingTitle, rounds]
              properties:
                openingTitle:
                  type: string
                  minLength: 1
                  maxLength: 120
                rounds:
                  type: array
                  maxItems: 100
                  items:
                    type: object
                loveLetter:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  description: love-journey@1.1.0 必填。
                letterPassword:
                  type: string
                  pattern: "^[0-9]{4}$"
                  description: love-journey@1.1.0 必填的四位数字拆信密码。
                passwordHint:
                  type: string
                  maxLength: 100
            assets:
              type: array
              items:
                $ref: "#/components/schemas/PublicAsset"
            playSessionExpiresAt:
              type: string
              format: date-time
        requestId:
          type: string
...
    PublicAssetListResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          type: object
          required: [assets]
          properties:
            assets:
              type: array
              items:
                $ref: "#/components/schemas/PublicAsset"
        requestId:
          type: string
...
    ErrorResponse:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
        requestId:
          type: string
    ULID:
      type: string
      pattern: "^[0-7][0-9A-HJKMNP-TV-Z]{25}$"
      minLength: 26
      maxLength: 26
...
    PublicPlayAnalyticsEventRequest:
      type: object
      additionalProperties: false
      required: [eventName, clientEventId, properties]
      properties:
        eventName:
          type: string
          enum: [play.completed, play.replayed]
        clientEventId:
          type: string
          format: uuid
          pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
        occurredAt:
          type: string
          format: date-time
        properties:
          $ref: "#/components/schemas/PublicPlayProperties"
...
    PublicPlayProperties:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode:
          type: string
          const: public
    AnalyticsEventWriteResponse:
      type: object
      additionalProperties: false
      required: [data, requestId]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [eventId, duplicate]
          properties:
            eventId:
              $ref: "#/components/schemas/ULID"
            duplicate:
              type: boolean
        requestId:
          $ref: "#/components/schemas/ULID"
...